@@ -1,7 +1,7 @@
 package com.ecommerce.dto;
 
-import javax.validation.constraints.Min;
-import javax.validation.constraints.NotNull;
+import jakarta.validation.constraints.Min;
+import jakarta.validation.constraints.NotNull;
 
 /**
  * DTO for cart item operations (add/update).
