@@ -1,14 +1,11 @@
 package com.ecommerce;
 
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
 
 /**
  * Main application context test.
  */
-@RunWith(SpringRunner.class)
 @SpringBootTest
 public class EcommerceApplicationTests {
 
