@@ -3,26 +3,26 @@ package com.ecommerce.service;
 import com.ecommerce.dto.ProductDto;
 import com.ecommerce.model.Product;
 import com.ecommerce.repository.ProductRepository;
-import org.junit.Before;
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
-import org.mockito.junit.MockitoJUnitRunner;
+import org.mockito.junit.jupiter.MockitoExtension;
 
 import java.math.BigDecimal;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Optional;
 
-import static org.junit.Assert.*;
+import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.*;
 
 /**
  * Unit tests for ProductService.
  */
-@RunWith(MockitoJUnitRunner.class)
+@ExtendWith(MockitoExtension.class)
 public class ProductServiceTest {
 
     @Mock
@@ -33,7 +33,7 @@ public class ProductServiceTest {
 
     private Product testProduct;
 
-    @Before
+    @BeforeEach
     public void setUp() {
         testProduct = new Product();
         testProduct.setProductId(1L);
@@ -139,14 +139,16 @@ public class ProductServiceTest {
         verify(productRepository, times(1)).save(any(Product.class));
     }
 
-    @Test(expected = RuntimeException.class)
+    @Test
     public void testUpdateProduct_NotFound() {
         ProductDto productDto = new ProductDto();
         productDto.setName("Updated Product");
 
         when(productRepository.findById(999L)).thenReturn(Optional.empty());
 
-        productService.updateProduct(999L, productDto);
+        assertThrows(RuntimeException.class, () -> {
+            productService.updateProduct(999L, productDto);
+        });
     }
 
     @Test
@@ -160,11 +162,13 @@ public class ProductServiceTest {
         verify(productRepository, times(1)).deleteById(1L);
     }
 
-    @Test(expected = RuntimeException.class)
+    @Test
     public void testDeleteProduct_NotFound() {
         when(productRepository.existsById(999L)).thenReturn(false);
 
-        productService.deleteProduct(999L);
+        assertThrows(RuntimeException.class, () -> {
+            productService.deleteProduct(999L);
+        });
     }
 
     @Test
@@ -178,11 +182,13 @@ public class ProductServiceTest {
         verify(productRepository, times(1)).save(any(Product.class));
     }
 
-    @Test(expected = RuntimeException.class)
+    @Test
     public void testReduceStock_InsufficientStock() {
         when(productRepository.findById(1L)).thenReturn(Optional.of(testProduct));
 
-        productService.reduceStock(1L, 150); // More than available stock
+        assertThrows(RuntimeException.class, () -> {
+            productService.reduceStock(1L, 150); // More than available stock
+        });
     }
 
     @Test
